Remove FAQ about limited duplicates
The rule about duplicates counting as your Limited card was changed in
FAQ v2.3 with the rule directly under the one being removed. While the
title references not allowing duplicated limited cards during setup,
none of the text references set up.

Since the two FAQs are contradictory, and the last ruling was that
duplicates do not count for Limited, remove the earlier FAQ.